requirements.txt
PyGithub==1.46
fabric==2.5.0